/v1/subscriptions: Exception 처리 추가 /v1/users/message: expiry 초단위로 수정

+++ domain/src/main/java/kr/co/adflow/push/domain/ktp/request/UserMessage.java
@@ -26,7 +26,7 @@ public class UserMessage {
 	private int qos;
 	
 	/** The expiry. */
-	private long expiry;
+	private int expiry;
 	//private int type;
 	
 	/**
@@ -124,7 +124,7 @@ public class UserMessage {
 	 *
 	 * @return the expiry
 	 */
-	public long getExpiry() {
+	public int getExpiry() {
 		return expiry;
 	}
 	
@@ -133,7 +133,7 @@ public class UserMessage {
 	 *
 	 * @param expiry the new expiry
 	 */
-	public void setExpiry(long expiry) {
+	public void setExpiry(int expiry) {
 		this.expiry = expiry;
 	}
